+++ src/UglyToad.PdfPig/CrossReference/CrossReferenceTablePart.cs
@@ -15,11 +15,11 @@
     ///     ggggg is a 5-digit generation number
     ///     n is a literal keyword identifying this as an in-use entry
     ///     eol is a 2-character end-of-line sequence
-    /// 
-    /// 
+    ///
+    ///
     /// The byte offset is a 10-digit number, padded with leading zeros if necessary,
     /// giving the number of bytes from the beginning of the file to the beginning of the
-    /// object. 
+    /// object.
     /// </remarks>
     internal class CrossReferenceTablePart
     {
@@ -56,7 +56,7 @@
         public void FixOffset(long offset)
         {
             Offset = offset;
-            Dictionary = Dictionary.With(NameToken.Prev, new NumericToken(offset));
+            Dictionary = Dictionary.With(NameToken.Prev, new NumericToken((double)offset));
         }
 
         public long GetPreviousOffset()
