+++ src/UglyToad.PdfPig/Graphics/Operations/General/SetLineWidth.cs
@@ -19,13 +19,13 @@
         /// <summary>
         /// The line width.
         /// </summary>
-        public decimal Width { get; }
+        public double Width { get; }
 
         /// <summary>
         /// Create a new <see cref="SetLineWidth"/>.
         /// </summary>
         /// <param name="width">The line width.</param>
-        public SetLineWidth(decimal width)
+        public SetLineWidth(double width)
         {
             Width = width;
         }
